+++ packages/datasets-fs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prosopo/datasets-fs
 
+## 3.0.34
+### Patch Changes
+
+- Updated dependencies [8f1773a]
+  - @prosopo/types@3.5.11
+
 ## 3.0.33
 ### Patch Changes