+++ packages/env/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @prosopo/env
 
+## 3.2.21
+### Patch Changes
+
+- Updated dependencies [8f1773a]
+  - @prosopo/types@3.5.11
+  - @prosopo/database@3.4.13
+  - @prosopo/keyring@2.8.35
+  - @prosopo/types-env@2.7.46
+
 ## 3.2.20
 ### Patch Changes