+++ packages/procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prosopo/procaptcha-wrapper
 
+## 2.6.37
+### Patch Changes
+
+- Updated dependencies [8f1773a]
+  - @prosopo/types@3.5.11
+
 ## 2.6.36
 ### Patch Changes