+++ packages/procaptcha/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @prosopo/procaptcha
 
+## 2.9.13
+### Patch Changes
+
+- Updated dependencies [8f1773a]
+  - @prosopo/types@3.5.11
+  - @prosopo/api@3.1.32
+  - @prosopo/datasets@3.0.42
+  - @prosopo/procaptcha-common@2.9.13
+
 ## 2.9.12
 ### Patch Changes